+++ CHANGES.rst
@@ -13,6 +13,12 @@ API changes
   - The ``find_peaks`` function will no longer issue a RuntimeWarning
     if the input data contains NaNs. [#712]
 
+  - ``DAOStarFinder`` and ``IRAFStarFinder`` gain two new parameters:
+    ``brightest`` to keep the top ``brightest`` (based on the flux) objects
+    in the returned catalog *after all other filtering has been applied* and
+    ``peakmax`` to exclude sources with peak pixel values larger or equal to
+    ``peakmax``. [#750]
+
 - ``photutils.epsf``
 
   - The ``Star``, ``Stars``, and ``LinkedStars`` classes are now